+++ CHANGELOG.md
@@ -3,12 +3,18 @@
 
 Any significant changes made to this project will be documented in this file.
 
-## [2.3.0] - 2025-12-09
+## [2.3.0] - 2025-12-23
 
 #### Added
 
-- #4 new supporting of image name checking.
 - checkingNameTypes set with cases: .firstUpperCase, .camelCase, .sneak_case, .kebab_case, .custom
+- #4 New supporting of image name checking.
+- #5 Filter for image name checking with a file size and a image size conditons.
+- Tests for checking Settings parser.
+
+#### Fixed
+
+- Issue with parsing Settings yaml file.
 
 ## [2.2.1] - 2025-11-29